Refactor CaptureStdOut into util
