+++ pkg/testcommon/testcommon_test.go
@@ -9,7 +9,6 @@ import (
 
 	"github.com/drud/ddev/pkg/ddevapp"
 	"github.com/drud/ddev/pkg/fileutil"
-	"github.com/drud/ddev/pkg/output"
 	"github.com/drud/ddev/pkg/util"
 	asrt "github.com/stretchr/testify/assert"
 )
@@ -73,28 +72,6 @@ func TestChdir(t *testing.T) {
 	assert.Equal(currentDir, startingDir, "Ensure we have changed back to the starting directory")
 
 	CleanupDir(testDir)
-}
-
-// TestCaptureUserOut ensures capturing of stdout works as expected.
-func TestCaptureUserOut(t *testing.T) {
-	assert := asrt.New(t)
-	restoreOutput := CaptureUserOut()
-	text := util.RandString(128)
-	output.UserOut.Println(text)
-	out := restoreOutput()
-
-	assert.Contains(out, text)
-}
-
-// TestCaptureStdOut ensures capturing of stdout works as expected.
-func TestCaptureStdOut(t *testing.T) {
-	assert := asrt.New(t)
-	restoreOutput := CaptureStdOut()
-	text := util.RandString(128)
-	fmt.Println(text)
-	out := restoreOutput()
-
-	assert.Contains(out, text)
 }
 
 // TestValidTestSite tests the TestSite struct behavior in the case of a valid configuration.
